+++ release_notes.md
@@ -10,3 +10,4 @@
   - [Added net9 prelaunch app.](https://github.com/Azure/azure-functions-dotnet-worker/pull/2898)
 - Update the `DefaultHttpProxyService` to better handle client disconnect scenarios (#10688)
   - Replaced `InvalidOperationException` with `HttpForwardingException` when there is a ForwarderError
+- Updated `WebJobs.Script` to target .NET 8 (instead of .NET Standard 2.1)